app/requirements.txt
anyio==3.6.1
asgiref==3.5.2
boto3==1.24.2
botocore==1.27.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
jmespath==1.0.0
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
